+++ inbm/Changelog.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - YYYY-MM-DD
+### Changed
+ - (NEX-15262) Returns failure reason that matches the format required by MM
 
 ## 4.2.7 - 2024-11-8
 ### Changed