+++ inbm/Changelog.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## NEXT - YYYY-MM-DD
+## Next Release
+### Changed
+ - (NEXMANAGE-513) Record the granular log in download-only mode
 
 ## 4.2.7 - 2024-11-8
 ### Changed