+++ inbm/Changelog.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ### Added
  - (NEXMANAGE-950) Add Set Power State Capability to cloudadapter from UDM
+ - (NEXMANAGE-826) Update granular log to support architecture independent package
 
 ### Fixed
  - (NEXMANAGE-733) Fixed typo in package_list during conversion from INBS to cloudadapter