+++ inbm/Changelog.md
@@ -3,8 +3,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## NEXT - YYYY-MM-DD
-## Next Release
+## 4.2.8 - 2024-11-29
 ### Changed
  - (NEXMANAGE-513) Record the granular log in download-only mode
  - (NEXMANAGE-1022) Adds Power Capabilities to Static Telemetry data.  Sends Static Telemetry data to INBS via the INBS Cloud Client in CloudAdapter.